# fix: type instability causing performance issue.